Add deserializing with serde
