+++ .ra.config.json
@@ -1,0 +1,7 @@
+{
+	"rust-analyzer":  {
+		"cargo": {
+			"features": ["std-write"] 
+		}
+	}
+}
